+++ NEWS.md
@@ -7,6 +7,8 @@
 * `glimpse()` makes it possible to see all the columns in a tbl,
   displaying as much data for each variable as can be fit on a single line.
 
+* `rbind_all()` is stricter and only accepts list of data frames (#288)  
+  
 ## Bug fixes
 
 * Code adapted to Rcpp > 0.11.1